jk actually renamed the files using git instead
