+++ src/components/Admin/Admins.jsx
@@ -4,7 +4,7 @@ import Title from "./Title.jsx";
 import Filters from "./Filters.jsx";
 import Toolbar from "./Toolbar.jsx";
 import Table from "./Table.jsx";
-import { adminList } from "@/data/mock/admin.js";
+import { adminList } from "@/data/mock/Admin.js";
 import { adminFilters } from "@/data/Filters.js";
 import { adminHeaders } from "@/data/Headers.js";
 import { adminTags } from "@/data/Tags.js";
