+++ src/components/Admin/CheckIn.jsx
@@ -2,12 +2,12 @@ import Title from "./Title";
 import ScanQRCode from "./ScanQRCode";
 import { useState } from "react";
 import DropDown from "./DropDown";
-import { mockEvents } from "@/data/mock/events";
+import { mockEvents } from "@/data/mock/Events";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
 import Button from "./Button";
 import CheckInfo from "./CheckInfo";
-import { checkInUser } from "@/data/mock/checkInUser";
+import { checkInUser } from "@/data/mock/CheckInUser";
 
 const CheckIn = () => {
   const setResult = (result) => {
