+++ src/components/Admin/Judges.jsx
@@ -4,7 +4,7 @@ import Title from "./Title.jsx";
 import Filters from "./Filters.jsx";
 import Toolbar from "./Toolbar.jsx";
 import Table from "./Table.jsx";
-import { judgeList } from "@/data/mock/judges.js";
+import { judgeList } from "@/data/mock/Judges.js";
 import { judgeFilters } from "@/data/Filters.js";
 import { judgeHeaders } from "@/data/Headers.js";
 import { judgeTags } from "@/data/Tags.js";
