+++ src/components/Admin/Mentors.jsx
@@ -5,7 +5,7 @@ import Filters from "@/components/Admin/Filters";
 import Toolbar from "@/components/Admin/Toolbar";
 import Title from "./Title";
 import Table from "./Table";
-import { mentorList } from "@/data/mock/mentors";
+import { mentorList } from "@/data/mock/Mentors";
 import { mentorTags } from "@/data/Tags.js";
 import { mentorFilters } from "@/data/Filters.js";
 import { mentorHeaders } from "@/data/Headers.js";
